Convert infrastructure to utilize bootstrapper
This effectively creates one source of truth on file and module resolution, preventing potential issues where module a depends on {b,c} and b depends on {c} and both hardcode different versions, leading to indeterminate behavior at runtime.

Additionally moves all of the loading logic into the initial entry point so there is minimal duplication across different components in the system.

Users are now expected to either load the bootstrap.js file in their site or to utilize the `output.bootstrap` flag which will inline the bootstrap logic into the main entry point. For cases where this is inlined, no other apps can link to this code as there are sanity checks in place to prevent multiple circus bootstraps being loading on the same page as they will conflict on a few global variables.

This has the added bonus of allowing us to rewrite dependencies of files at publish time, allowing us to defer the local vs. not decision across the entire app. (Those changes to be finalized in a future commit)