Add circus version build check
We want to invalidate components that are built using a different major version as we can not guarantee that they are compatible after a major version change. This forces us to fail early, at the expense of potentially having cascading rebuild/republish cycles should breaking changes need to occur within Circus itself.

Fixes #32