Prevent duplicate scripts prior to DOM ready
When loading from a response that has scripts predictively included, it’s possible to have multiple inserts for a given module as the DOM is not complete and we are unable to find the soon to be rendered scripts. By deferring slightly, we ensure that only one instance of a given library is loaded.